sw: fix stack overflow on kde45196-1.html
Previously, the shell cursor was put onto a SwGrfNode in the top left
corner, which is invalid, but didn't crash immediately.

With commit b20ca8d951e8205c8b963c6b7407f984053b4094 the cursor is
instead put onto the first SwTextNode in the body, which happens to be
in a table, and the entire table is hidden; this causes an infinite
recursion in SwCursorShell::UpdateCursorPos() because
GetModelPositionForViewPoint() no longer moves it to SwGrfNode.

So try to move the cursor out of a hidden node, which requires an
additional change in SwCursor::IsSelOvr() to allow moving from a node
without a frame to another node without a frame.

Reviewed-on: https://gerrit.libreoffice.org/c/core/+/158119
Tested-by: Michael Stahl <michael.stahl@allotropia.de>
(cherry picked from commit fa4fe53aa775cdc50eec715f4bd88469fe5e0d32)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/158048
Tested-by: Jenkins
(cherry picked from commit 92d75686181349f4b79aef413a070ab8335f97b9)

+++ sw/source/core/crsr/swcrsr.cxx
@@ -394,9 +394,16 @@ bool SwCursor::IsSelOvr(SwCursorSelOverFlags const eFlags)
 
         if( !pFrame )
         {
-            DeleteMark();
-            RestoreSavePos();
-            return true; // we need a frame
+            assert(!m_vSavePos.empty());
+            SwContentNode const*const pSaveNode(rNds[m_vSavePos.back().nNode]->GetContentNode());
+            // if the old position already didn't have a frame, allow moving
+            // anyway, hope the caller can handle that
+            if (pSaveNode && pSaveNode->getLayoutFrame(rDoc.getIDocumentLayoutAccess().GetCurrentLayout()))
+            {
+                DeleteMark();
+                RestoreSavePos();
+                return true; // we need a frame
+            }
         }
     }
